Migrate from MightyPirates to MovingBlocks JNLua [phase 1]

+++ src/main/scala/li/cil/oc/server/machine/luac/UnicodeAPI.scala
@@ -45,7 +45,7 @@ class UnicodeAPI(owner: NativeLuaArchitecture) extends NativeLuaAPI(owner) {
         })
         else string.length
       if (end <= start) lua.pushString("")
-      else lua.pushString(string.substring(start, end))
+      else lua.pushString(string.substring(start.toInt, end.toInt))
       1
     })
     lua.setField(-2, "sub")
